pyproject.toml
[project]
name = "voice-transcriber"
version = "0.1.0"
description = ""
authors = [
    {name = "johning",email = "52798743+johnt0@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "boto3 (>=1.38.8,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
